1.add the sortlist interface.2.fix the problem when using with the utils's rrupdate interface.

+++ dhcp/postgres/tables.go
@@ -8,7 +8,7 @@ type Lease4 struct {
 	Address       int
 	Hwaddr        []byte
 	ClientId      []byte
-	ValidLifetime int64
+	ValidLifetime int32
 	Expire        time.Time
 	SubnetId      int64
 	FqdnFwd       bool
